% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{insta_login}
\alias{insta_login}
\title{insta_login}
\usage{
insta_login(passwd = "", save = F, load = F)
}
\arguments{
\item{passwd}{Specify your Instagram password. If you leave empty a prompt will ask you for input.}

\item{save}{save your session so you don't need to specify password again}

\item{load}{load the instagram session by just setting INSTAGRAM_LOGIN environment variable. Will only work if you have saved your credentials before.}
}
\description{
insta_login
}
