% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_instaloadeR}
\alias{install_instaloadeR}
\title{install_instaloadeR}
\usage{
install_instaloadeR(envname = NULL)
}
\arguments{
\item{envname}{specify Python environment name for module installation}
}
\description{
install_instaloadeR
}
